requirements.txt
# FastAPI & Server
fastapi==0.115.6
pydantic[email]==2.12.5
uvicorn[standard]==0.32.1

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2   # ✅ pinned for passlib compatibility

# Database
sqlalchemy==2.0.36

# ML & Image
numpy==2.2.1
scikit-learn==1.6.0
Pillow==11.0.0

# HTTP
python-multipart==0.0.18
httpx==0.28.1

# LLM
openai==1.57.2

# Environment
python-dotenv==1.0.1
